[gym_jiminy/common] Move leaky twist from mahony filter to body observer.
